backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
Flask-Mail==0.10.0
stripe==11.1.1
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.4
